Modified project tree, fixed most arch independence V core code

+++ kernel/include/core/debug.h
@@ -4,8 +4,8 @@
 #include <stdint.h>
 
 void debug_print_str(const char* str);
-void debug_print(uint32_t value) ;
-void debug_print_int_nonewline(int32_t value);
-void debug_print_int(int32_t value) ;
+void debug_print(uint64_t value) ;
+void debug_print_int_nonewline(int64_t value);
+void debug_print_int(int64_t value) ;
 
 #endif // __DEBUG_H__
